Relocate protocol implementation to meet package structure.
