+++ protocol/src/main/java/de/d3adspace/actuarius/protocol/DNSMessage.java
@@ -22,10 +22,10 @@
  * SOFTWARE.
  */
 
-package de.d3adspace.actuarius.server.protocol;
+package de.d3adspace.actuarius.protocol;
 
 /**
  * @author Felix Klauke <fklauke@itemis.de>
  */
-public interface DNSResponse {
+public interface DNSMessage {
 }
