+++ protocol/src/main/java/de/d3adspace/actuarius/protocol/DNSQuery.java
@@ -22,10 +22,12 @@
  * SOFTWARE.
  */
 
-package de.d3adspace.actuarius.server.protocol;
+package de.d3adspace.actuarius.protocol;
 
 /**
  * @author Felix Klauke <fklauke@itemis.de>
  */
-public interface DNSMessage {
+public interface DNSQuery {
+
+
 }
