+++ server/src/main/java/de/d3adspace/actuarius/protocol/DNSOperationCode.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package de.d3adspace.actuarius.server.protocol;
+package de.d3adspace.actuarius.protocol;
 
 /**
  * @author Felix Klauke <fklauke@itemis.de>
